* merge nnc linear

* modify UT

* add linear+log & linear+round

* remove bias maybe owned

* remove risky op

* turn on fast path

* merge nnc remain parts

* fix bug

* polish UT

* add flag to turn off fast path

* add fallback path and remove inplaced fusers

* Fix fast path memory layout mix-up issue

* fix bugs

* merge nnc linear

* modify UT

* add linear+log & linear+round

* remove risky op

* merge nnc remain parts

* add fallback path and remove inplaced fusers

* fix

* turn off fast pass flag

* disable UT

+++ csrc/jit/cpu/tensorexpr/external_call/conv_pow.h
@@ -52,7 +52,7 @@ struct LoweringFuncTrait<ConvFusedOp::kConvPow> : public ConvCommonOperations {
   static ideep::attr_t get_attr(int64_t* extra_args) {
     constexpr int exponent_idx = 0;
     const float exponent =
-        static_cast<float>(((int64_t*)extra_args)[exponent_idx]);
+        static_cast<float>(((double*)extra_args)[exponent_idx]);
     return ideep::attr_t::fuse_pow(1.0, 1.0, exponent);
   }
 };
